requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==3.2.9
django-ipware==4.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
email-validator==1.1.3
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
